% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheBoot.R
\name{getTheBoot}
\alias{getTheBoot}
\title{Random Replace Bootstrap}
\usage{
getTheBoot(
  theDataK,
  thePoint,
  theC,
  RandomReplaceRatio = 0.5,
  maxBoot = 200,
  tempPrint = "",
  ifItem = FALSE
)
}
\arguments{
\item{theDataK}{The data matrix of factor K.}

\item{thePoint}{The expectation matrix of factor K.}

\item{theC}{The category of factor K.}

\item{RandomReplaceRatio}{Decimal between 0 and 1.}

\item{maxBoot}{The maximum steps of bootstrap.}

\item{tempPrint}{= ''}

\item{ifItem}{= FALSE, calculate the items when TRUE.}
}
\value{
The function returns an array.
}
\description{
The bootstrap of random replacement.
}
\examples{
\dontrun{
NA
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
