% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheCategory.R
\name{getTheCategory}
\alias{getTheCategory}
\title{Calculate the Category}
\usage{
getTheCategory(theData)
}
\arguments{
\item{theData}{The prepared data matrix.}
}
\value{
The function returns an array.
}
\description{
Calculate the category of the data.
}
\examples{
\dontrun{
theCategory <- getTheCategory(theData)
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
