% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheData.R
\name{getTheData}
\alias{getTheData}
\title{Prepare Data for Analysis}
\usage{
getTheData(theTempData, ifInf = "edge", ifNA = "central", k = 10)
}
\arguments{
\item{theTempData}{The data matrix to analysis.}

\item{ifInf}{= 'NA' 'edge'}

\item{ifNA}{= 'omit' 'knn' 'central'}

\item{k}{= 10}
}
\value{
The function returns an array.
}
\description{
Transform the data matrix into the specific format for analysis.
}
\examples{
\dontrun{
theData <- getTheData(theTempData)
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
