% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheDeviation.R
\name{getTheDeviation}
\alias{getTheDeviation}
\title{Calculate the Deviations}
\usage{
getTheDeviation(
  theData,
  theExpectPoint,
  theCategory,
  theDim,
  RandomReplaceRatio = 0.5,
  maxBoot = 200,
  theSig = 0.05
)
}
\arguments{
\item{theData}{The prepared data matrix.}

\item{theExpectPoint}{The expectation matrix.}

\item{theCategory}{The Calculated Category.}

\item{theDim}{The Calculated Dimensions.}

\item{RandomReplaceRatio}{Decimal between 0 and 1.}

\item{maxBoot}{The maximum steps of bootstrap.}

\item{theSig}{= 0.05}
}
\value{
The function returns a list.
}
\description{
Calculate deviations between observations and expectations.
}
\examples{
\dontrun{
theDeviation <- getTheDeviation(theData, theExpectPoint, theCategory, theDim)
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
