% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheDim.R
\name{getTheDim}
\alias{getTheDim}
\title{Calculate the Dimensions}
\usage{
getTheDim(theData, center = TRUE, scale = TRUE, fm = "mle", rotate = "varimax")
}
\arguments{
\item{theData}{The prepared data matrix.}

\item{center}{= TRUE}

\item{scale}{= TRUE}

\item{fm}{= 'mle'}

\item{rotate}{= 'varimax'}
}
\value{
The function returns an array.
}
\description{
Calculate the dimensions of the data by factor analysis. Factors for less than two items will be combined
}
\examples{
\dontrun{
theDim <- getTheDim(theData)
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
