% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheExpectPoint.R
\name{getTheExpectPoint}
\alias{getTheExpectPoint}
\title{Estimate the Expectations}
\usage{
getTheExpectPoint(
  theData,
  theCategory,
  theDim,
  theModel = "GGUM",
  pack = "GGUM",
  SE = FALSE,
  precision = 4,
  N_nodes = 30,
  max_outer = 60,
  max_inner = 60,
  tol = 0.001,
  method = "EM"
)
}
\arguments{
\item{theData}{The prepared data matrix.}

\item{theCategory}{The Calculated Category.}

\item{theDim}{The Calculated Dimensions.}

\item{theModel}{= 'GGUM' 'GPCM' 'PCM'}

\item{pack}{= 'GGUM' 'mirt'}

\item{SE}{= FALSE}

\item{precision}{= 4}

\item{N_nodes}{= 30}

\item{max_outer}{= 60}

\item{max_inner}{= 60}

\item{tol}{= 0.001}

\item{method}{= 'EM'}
}
\value{
The function returns a list.
}
\description{
Estimate the expectations of each person on each item.
}
\examples{
\dontrun{
theExpectPoint <- getTheExpectPoint(theData, theCategory, theDim)
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
