% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheGGUM.R
\name{getTheGGUM}
\alias{getTheGGUM}
\title{Estimate by GGUM}
\usage{
getTheGGUM(
  theDataK,
  theC,
  pack = "GGUM",
  SE = FALSE,
  precision = 4,
  N_nodes = 30,
  max_outer = 60,
  max_inner = 60,
  tol = 0.001,
  method = "EM"
)
}
\arguments{
\item{theDataK}{The data matrix of factor K.}

\item{theC}{The category of factor K.}

\item{pack}{= 'GGUM' 'mirt'}

\item{SE}{= FALSE}

\item{precision}{= 4}

\item{N_nodes}{= 30}

\item{max_outer}{= 60}

\item{max_inner}{= 60}

\item{tol}{= 0.001}

\item{method}{= 'EM'}
}
\value{
The function returns a list.
}
\description{
Get the estimation of the GGUM.
}
\examples{
\dontrun{
NA
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
