% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheGPCM.R
\name{getTheGPCM}
\alias{getTheGPCM}
\title{Estimate by GPCM}
\usage{
getTheGPCM(theDataK, theC, SE = FALSE, method = "EM")
}
\arguments{
\item{theDataK}{The data matrix of factor K.}

\item{theC}{The category of factor K.}

\item{SE}{= FALSE}

\item{method}{= 'EM'}
}
\value{
The function returns an array.
}
\description{
Get the estimation of the GPCM.
}
\examples{
\dontrun{
NA
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
