% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheGRM.R
\name{getTheGRM}
\alias{getTheGRM}
\title{Estimate by GRM}
\usage{
getTheGRM(theDataK, theC, SE = FALSE)
}
\arguments{
\item{theDataK}{The data matrix of factor K.}

\item{theC}{The category of factor K.}

\item{SE}{= FALSE}
}
\value{
The function returns an array.
}
\description{
Get the estimation of the GRM.
}
\examples{
\dontrun{
NA
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
