% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getThePCM.R
\name{getThePCM}
\alias{getThePCM}
\title{Estimate by PCM}
\usage{
getThePCM(theDataK, theC, SE = FALSE)
}
\arguments{
\item{theDataK}{The data matrix of factor K.}

\item{theC}{The category of factor K.}

\item{SE}{= FALSE}
}
\value{
The function returns an array.
}
\description{
Get the estimation of the PCM.
}
\examples{
\dontrun{
NA
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
