% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getThePercent.R
\name{getThePercent}
\alias{getThePercent}
\title{Step Percent of Random Replacement Deviation}
\usage{
getThePercent(
  theData,
  theExpectPoint,
  theCategory,
  theDim,
  theStep = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1),
  maxBoot = 1000,
  theCompare = "upper",
  theAlt = "less",
  theSig = 0.05
)
}
\arguments{
\item{theData}{The prepared data matrix.}

\item{theExpectPoint}{The expectation matrix.}

\item{theCategory}{The Calculated Category.}

\item{theDim}{The Calculated Dimensions.}

\item{theStep}{An array of random replacement ratio.}

\item{maxBoot}{The maximum steps of bootstrap.}

\item{theCompare}{= 'between', 'lower', 'upper', 'mean'}

\item{theAlt}{= 'two.sided', 'greater', 'less'}

\item{theSig}{= 0.05}
}
\value{
The function returns a list.
}
\description{
Step by step list the proportionally random expectation deviations.
}
\examples{
\dontrun{
thePercent <- getThePercent(theData, theExpectPoint, theCategory, theDim)
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
