% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NERF.R
\name{NERF}
\alias{NERF}
\title{Non-Effortful Responses Filter}
\usage{
NERF(
  theTempData,
  ifInf = "edge",
  ifNA = "central",
  k = 10,
  center = TRUE,
  scale = TRUE,
  fm = "mle",
  rotate = "varimax",
  theModel = "GPCM",
  pack = "mirt",
  SE = FALSE,
  precision = 4,
  N_nodes = 30,
  max_outer = 60,
  max_inner = 60,
  tol = 0.001,
  method = "EM",
  RandomReplaceRatio = 0.5,
  maxBoot = 200,
  theSig = 0.05
)
}
\arguments{
\item{theTempData}{The data matrix to analysis.}

\item{ifInf}{= 'NA' 'edge'}

\item{ifNA}{= 'omit' 'knn' 'central'}

\item{k}{= 10}

\item{center}{= TRUE}

\item{scale}{= TRUE}

\item{fm}{= 'mle'}

\item{rotate}{= 'varimax'}

\item{theModel}{= 'GGUM' 'GPCM' 'PCM'}

\item{pack}{= 'GGUM' 'mirt'}

\item{SE}{= FALSE}

\item{precision}{= 4}

\item{N_nodes}{= 30}

\item{max_outer}{= 60}

\item{max_inner}{= 60}

\item{tol}{= 0.001}

\item{method}{= 'EM'}

\item{RandomReplaceRatio}{Decimal between 0 and 1.}

\item{maxBoot}{The maximum steps of bootstrap.}

\item{theSig}{= 0.05}
}
\value{
The function returns a list.
}
\description{
Detect non-effortful responses in questionnaires or examinations.
}
\examples{
\dontrun{
theDeviation <- NERF(theTempData)
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
