% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NERF.R
\name{NERF}
\alias{NERF}
\title{Non-Effortful Responses Filter}
\usage{
NERF(
  theTempData,
  ifInf = "edge",
  ifNA = "central",
  k = 10,
  center = TRUE,
  scale = TRUE,
  fm = "mle",
  rotate = "varimax",
  theModel = "GPCM",
  pack = "mirt",
  SE = FALSE,
  precision = 4,
  N_nodes = 30,
  max_outer = 60,
  max_inner = 60,
  tol = 0.001,
  method = "EM",
  theStep = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1),
  maxBoot = 1000,
  theCompare = "upper",
  theAlt = "less",
  theSig = 0.05
)
}
\arguments{
\item{theTempData}{The data matrix to analysis.}

\item{ifInf}{= 'NA' 'edge'}

\item{ifNA}{= 'omit' 'knn' 'central'}

\item{k}{= 10}

\item{center}{= TRUE}

\item{scale}{= TRUE}

\item{fm}{= 'mle'}

\item{rotate}{= 'varimax'}

\item{theModel}{= 'GGUM' 'GPCM' 'PCM' 'GRM'}

\item{pack}{= 'GGUM' 'mirt'}

\item{SE}{= FALSE}

\item{precision}{= 4}

\item{N_nodes}{= 30}

\item{max_outer}{= 60}

\item{max_inner}{= 60}

\item{tol}{= 0.001}

\item{method}{= 'EM'}

\item{theStep}{An array of random replacement ratio.}

\item{maxBoot}{The maximum steps of bootstrap.}

\item{theCompare}{= 'between', 'lower', 'upper', 'mean'}

\item{theAlt}{= 'two.sided', 'greater', 'less'}

\item{theSig}{= 0.05}
}
\value{
The function returns a list.
}
\description{
Detect non-effortful responses in questionnaires or examinations.
}
\examples{
\dontrun{
theNERF <- NERF(theTempData)
}

}
\author{
zdx, \email{zhaodexuan@aliyun.com}
}
